requirements.txt
# Basic functionalies
opencv-python
tensorflow==2.15
keras
pillow==10.2.0
numpy

# openvino and its dependencies
openvino-dev>=2024.0.0
openvino-telemetry>=2023.2.1
nncf==2.9.0
openvino-dev[onnx]

# jupyter 
jupyterlab
ipywidgets
ipykernel>=5.*
ipython>=7.16.3


# Flask and related dependencies
Flask
Flask-Cors
Flask-Caching
Flask-Logging

# Others
numpy>=1.21.0
Pillow>=8.3.2
matplotlib>=3.4,<3.7; platform_system == "Windows"
matplotlib>=3.4; platform_system != "Windows"
scipy
silence-tensorflow
